# docker/openapi.yaml
openapi: 3.0.3
...
info:
  title: Gestion de Bibliothèque — API
  version: 0.0.1-SNAPSHOT
  description: |
    API REST du projet bibliothèque (Spring Boot 2.4.5).

    ## Comptes disponibles

    Mot de passe `admin123` pour tous.

    | Identifiant | Rôle | Rôle dans le scénario de test |
    |---|---|---|
    | `admin` | Admin | administration des livres et des utilisateurs |
    | `a1` | User | **A1**, adhérent réservataire — `adherentId` = 2 |
    | `a2` | User | **A2**, celui qui saturera son quota (RG-03) — `adherentId` = 3 |
    | `a3` | User | **A3**, l'emprunteur : il détient L2 à L5 — `adherentId` = 4 |

    ## Livres du scénario de test

    | Réf. | `livreId` | État |
    |---|---|---|
    | **L1** | 1 | disponible, 3 exemplaires — le réserver déclenche RG-01 |
    | **L2** à **L5** | 2, 3, 4, 5 | empruntés par A3 et non rendus — réservables |
servers:
  - url: http://localhost:8080
    description: Pile docker compose locale
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Jeton renvoyé par POST /authenticate. Ajouté automatiquement par
        auth.interceptor.ts côté Angular, et lu ici par JwtRequestFilter.
  schemas:
...
    StatutReservation:
      type: string
      enum: [EN_ATTENTE, DISPONIBLE, ANNULEE, EXPIREE, HONOREE]
      description: >
        EN_ATTENTE et DISPONIBLE sont les statuts « actifs » : eux seuls
        comptent pour RG-02 et RG-03, et eux seuls autorisent une annulation
        (RG-05). ANNULEE, EXPIREE et HONOREE sont définitifs (RG-06).
      example: EN_ATTENTE
    ReservationRequest:
      type: object
      description: >
        Le corps accepté par POST /api/reservations. Deux champs, pas un de
        plus : les dates et le statut sont posés par le serveur. Envoyer un
        dateExpiration n'aurait aucun effet — le champ est simplement ignoré.


        Depuis la séance 4, adherentId n'est plus une source de vérité
        (RS-04) : l'identité vient du jeton. Un ADHERENT peut l'omettre ; s'il
        le renseigne, ce doit être le sien, sinon 403. Seul un BIBLIOTHECAIRE,
        qui réserve au nom d'un adhérent, doit le fournir.
      required: [livreId]
      properties:
        livreId:
          type: integer
          description: bookId d'un livre dont noOfCopies vaut 0 (RG-01).
          example: 2
        adherentId:
          type: integer
          description: >
            userId de l'adhérent pour qui réserver. Obligatoire pour un
            BIBLIOTHECAIRE ; facultatif pour un ADHERENT, et forcément le sien.
          example: 2
    ReservationResponse:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: >
            Colonne serial PostgreSQL, alimentée par la séquence
            reservation_reservation_id_seq, distincte de hibernate_sequence qui
            numérote les livres et les utilisateurs. Le jeu de test démarre
            ce compteur à 100, pour qu'un identifiant de réservation ne puisse
            pas être confondu avec un livreId ou un adherentId.
          example: 100
        livreId:
          type: integer
          example: 2
        livreTitre:
          type: string
          example: L2 — L'Étranger
        adherentId:
          type: integer
          example: 2
        adherentNom:
          type: string
          description: >
            Le nom, et rien d'autre. Contrairement à /admin/users, aucun haché
            de mot de passe ne transite : c'est tout l'intérêt du DTO.
          example: A1 Réservataire principal
        dateReservation:
          type: string
          format: date-time
          readOnly: true
          description: Horodatage serveur, ISO-8601.
          example: '2026-08-21T10:32:07.412'
        dateExpiration:
          type: string
          format: date-time
          readOnly: true
          description: 'RG-04 : dateReservation + 7 jours.'
          example: '2026-08-28T10:32:07.412'
        statut:
          $ref: '#/components/schemas/StatutReservation'
    ApiError:
      type: object
      description: >
        Corps d'erreur du module Réservation. Un @ResponseStatus seul aurait
        renvoyé le bon code avec un corps muet (server.error.include-message
        vaut « never » depuis Spring Boot 2.3) : l'appelant aurait lu « 409 »
        sans savoir quelle règle il enfreint.
      properties:
        horodatage:
          type: string
          format: date-time
          example: '2026-08-21T10:32:07.412'
        statut:
          type: integer
          example: 409
        erreur:
          type: string
          example: Conflict
        message:
          type: string
          example: 'RG-01 : le livre « L1 — Le Petit Prince » est disponible (3 exemplaire(s) en rayon) ; il doit être emprunté, pas réservé.'
paths:
...
  /api/reservations:
    post:
      tags: [Réservations]
      summary: Créer une réservation
      description: >
        Le client n'envoie que livreId et adherentId. dateReservation,
        dateExpiration et statut sont déterminés par le serveur.


        ADHERENT : réserve pour lui-même uniquement, l'identité vient du jeton
        (RS-04). BIBLIOTHECAIRE : réserve au nom de n'importe quel adhérent.


        Ordre des contrôles : identité (403), champs obligatoires (400),
        existence du livre et de l'adhérent (404), puis règles de gestion
        (409). Un identifiant inconnu est donc signalé comme tel avant qu'on
        ne se prononce sur la disponibilité du livre.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: >
            Réservation créée, au statut EN_ATTENTE. L'en-tête Location pointe
            sur la ressource.
          headers:
            Location:
              description: URL de la réservation créée
              schema:
                type: string
                example: http://localhost:8080/api/reservations/100
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          description: >
            livreId ou adherentId absent. Le message nomme le champ manquant :
            « Le champ « livreId » est obligatoire. »
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 400
                erreur: Bad Request
                message: 'Le champ « livreId » est obligatoire.'
        '401':
          description: >
            RS-01 — jeton absent, expiré ou invalide. Le message précise
            lequel ; un jeton expiré invite à se reconnecter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 401
                erreur: Unauthorized
                message: 'Le jeton a expiré : reconnectez-vous via POST /authenticate.'
        '403':
          description: >
            RS-04 — un ADHERENT a envoyé l'adherentId d'un autre adhérent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 403
                erreur: Forbidden
                message: 'RS-04 : un adhérent ne réserve qu''en son nom ; adherentId=3 n''est pas le vôtre (2).'
        '404':
          description: Aucun livre, ou aucun adhérent, avec cet identifiant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 404
                erreur: Not Found
                message: "Aucun livre avec l'identifiant 987654."
        '409':
          description: >
            Règle de gestion enfreinte. Le message commence par la référence de
            la règle : RG-01 (livre disponible), RG-02 (doublon actif) ou RG-03
            (plus de trois réservations actives).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                RG-01:
                  value:
                    horodatage: '2026-08-21T10:32:07.412'
                    statut: 409
                    erreur: Conflict
                    message: 'RG-01 : le livre « L1 — Le Petit Prince » est disponible (3 exemplaire(s) en rayon) ; il doit être emprunté, pas réservé.'
                RG-02:
                  value:
                    horodatage: '2026-08-21T10:32:07.412'
                    statut: 409
                    erreur: Conflict
                    message: 'RG-02 : A1 Réservataire principal a déjà une réservation active sur « L2 — L''Étranger ».'
                RG-03:
                  value:
                    horodatage: '2026-08-21T10:32:07.412'
                    statut: 409
                    erreur: Conflict
                    message: 'RG-03 : A2 Quota à saturer a déjà 3 réservations actives ; le maximum est de 3.'

    get:
      tags: [Réservations]
      summary: Lister les réservations
      description: >
        Les deux filtres sont facultatifs et cumulables. Un adherentId inconnu
        ne déclenche pas de 404 : un filtre sans résultat renvoie une liste
        vide, ce qui est la réponse correcte à « montre-moi ses réservations ».


        ADHERENT : uniquement ses propres réservations, le filtre adherentId
        est forcé au sien (RS-05). BIBLIOTHECAIRE : toutes.
      parameters:
        - name: statut
          in: query
          required: false
          description: Filtre par statut. Insensible à la casse.
          schema:
            $ref: '#/components/schemas/StatutReservation'
        - name: adherentId
          in: query
          required: false
          description: >
            Filtre par adhérent (userId). Pour un ADHERENT, seul son propre
            identifiant est accepté ; un autre vaut 403.
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: Les réservations correspondant aux filtres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
        '400':
          description: >
            Statut inconnu. Le message liste les valeurs acceptées plutôt que
            de laisser échouer la conversion d'enum.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 400
                erreur: Bad Request
                message: 'Statut « PIZZA » inconnu. Valeurs acceptées : EN_ATTENTE, DISPONIBLE, ANNULEE, EXPIREE, HONOREE.'
        '401':
          description: >
            RS-01 — jeton absent, expiré ou invalide. Le message précise
            lequel ; un jeton expiré invite à se reconnecter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 401
                erreur: Unauthorized
                message: 'Le jeton a expiré : reconnectez-vous via POST /authenticate.'
        '403':
          description: >
            RS-05 — un ADHERENT a filtré sur l'adherentId d'un autre adhérent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 403
                erreur: Forbidden
                message: 'RS-05 : un adhérent ne consulte que ses propres réservations ; le filtre adherentId=3 n''est pas le vôtre.'

  # ---------------------------------------------------------------------------
  /api/reservations/expirees:
    get:
      tags: [Réservations]
      summary: Lister les réservations expirées
      description: >
        BONUS. Les réservations au statut EXPIREE.


        Le statut n'est pas déduit à l'affichage mais écrit en base par
        ReservationExpirationScheduler, qui balaie les réservations échues
        toutes les heures. Sinon une réservation périmée continuerait de peser
        dans RG-02 et RG-03, et bloquerait l'adhérent.


        Chemin littéral déclaré à côté de /{id} : Spring donne toujours la
        priorité au littéral sur le gabarit, « expirees » n'est donc jamais
        interprété comme un identifiant.


        Vue transversale sur tous les adhérents : réservée au BIBLIOTHECAIRE.
      responses:
        '200':
          description: Les réservations expirées
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: >
            RS-01 — jeton absent, expiré ou invalide. Le message précise
            lequel ; un jeton expiré invite à se reconnecter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 401
                erreur: Unauthorized
                message: 'Le jeton a expiré : reconnectez-vous via POST /authenticate.'
        '403':
          description: >
            RS-02 — le demandeur est un ADHERENT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 403
                erreur: Forbidden
                message: 'RS-02 : cette action est réservée au rôle BIBLIOTHECAIRE.'

  # ---------------------------------------------------------------------------
  /api/reservations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        example: 100

    get:
      tags: [Réservations]
      summary: Consulter une réservation
      description: >
        ADHERENT : seulement si elle lui appartient (RS-03).
        BIBLIOTHECAIRE : toutes.
      responses:
        '200':
          description: La réservation demandée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: >
            RS-01 — jeton absent, expiré ou invalide. Le message précise
            lequel ; un jeton expiré invite à se reconnecter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 401
                erreur: Unauthorized
                message: 'Le jeton a expiré : reconnectez-vous via POST /authenticate.'
        '403':
          description: >
            RS-03 — la réservation appartient à un autre adhérent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 403
                erreur: Forbidden
                message: 'RS-03 : la réservation 101 ne vous appartient pas.'
        '404':
          description: Aucune réservation avec cet identifiant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 404
                erreur: Not Found
                message: "Aucune réservation avec l'identifiant 987654."

    delete:
      tags: [Réservations]
      summary: Supprimer une réservation
      description: >
        Suppression définitive de la ligne, quel que soit son statut. À ne pas
        confondre avec l'annulation : annuler conserve la trace au statut
        ANNULEE, supprimer efface.


        Réservée au BIBLIOTHECAIRE (RS-02) : un ADHERENT reçoit 403, même sur
        sa propre réservation.
      responses:
        '204':
          description: Suppression effectuée, sans corps de réponse
        '401':
          description: >
            RS-01 — jeton absent, expiré ou invalide. Le message précise
            lequel ; un jeton expiré invite à se reconnecter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 401
                erreur: Unauthorized
                message: 'Le jeton a expiré : reconnectez-vous via POST /authenticate.'
        '403':
          description: >
            RS-02 — le demandeur est un ADHERENT.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 403
                erreur: Forbidden
                message: 'RS-02 : cette action est réservée au rôle BIBLIOTHECAIRE.'
        '404':
          description: Aucune réservation avec cet identifiant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 404
                erreur: Not Found
                message: "Aucune réservation avec l'identifiant 987654."

  # ---------------------------------------------------------------------------
  /api/reservations/{id}/annuler:
    patch:
      tags: [Réservations]
      summary: Annuler une réservation
      description: >
        Bascule le statut en ANNULEE. PATCH et non PUT : un seul champ change,
        la ressource n'est pas remplacée.


        RG-05 et RG-06 sont les deux faces du même contrôle : on n'annule que
        depuis EN_ATTENTE ou DISPONIBLE, donc jamais depuis un statut définitif.


        ADHERENT : seulement sa propre réservation (RS-03). BIBLIOTHECAIRE :
        toutes. Le contrôle de propriété précède celui du statut.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 100
      responses:
        '200':
          description: Réservation annulée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: >
            RS-01 — jeton absent, expiré ou invalide. Le message précise
            lequel ; un jeton expiré invite à se reconnecter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 401
                erreur: Unauthorized
                message: 'Le jeton a expiré : reconnectez-vous via POST /authenticate.'
        '403':
          description: >
            RS-03 — la réservation appartient à un autre adhérent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-09-12T21:40:00.000'
                statut: 403
                erreur: Forbidden
                message: 'RS-03 : la réservation 101 ne vous appartient pas.'
        '404':
          description: Aucune réservation avec cet identifiant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 404
                erreur: Not Found
                message: "Aucune réservation avec l'identifiant 987654."
        '409':
          description: 'RG-05 / RG-06 : la réservation est déjà dans un statut définitif.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                horodatage: '2026-08-21T10:32:07.412'
                statut: 409
                erreur: Conflict
                message: 'RG-05 : seule une réservation EN_ATTENTE ou DISPONIBLE peut être annulée ; celle-ci est ANNULEE. RG-06 : un statut définitif ne peut plus changer.'

# Toutes les routes exigent un jeton, sauf celles qui redéfinissent
# "security: []" ci-dessus.
security:
  - bearerAuth: []